requirements.txt
fastapi~=0.124.2
uvicorn[standard]
jinja2

databases
sqlalchemy
asyncpg

python-dotenv
psycopg2-binary